Improve DB seed script
